+++ nautobot_golden_config/nornir_plays/config_backup.py
@@ -17,6 +17,7 @@ from nautobot_plugin_nornir.utils import get_dispatcher
 
 from nautobot_golden_config.utilities.helper import (
     get_job_filter,
+    get_repository_working_dir,
     verify_global_settings,
     render_jinja_template,
 )
@@ -32,7 +33,7 @@ InventoryPluginRegister.register("nautobot-inventory", NautobotORMInventory)
 
 
 def run_backup(  # pylint: disable=too-many-arguments
-    task: Task, logger, global_settings, remove_regex_dict, replace_regex_dict, backup_root_folder
+    task: Task, logger, global_settings, remove_regex_dict, replace_regex_dict
 ) -> Result:
     r"""Backup configurations to disk.
 
@@ -54,8 +55,9 @@ def run_backup(  # pylint: disable=too-many-arguments
     backup_obj.backup_last_attempt_date = task.host.defaults.data["now"]
     backup_obj.save()
 
+    backup_directory = get_repository_working_dir("backup", obj, logger, global_settings)
     backup_path_template_obj = render_jinja_template(obj, logger, global_settings.backup_path_template)
-    backup_file = os.path.join(backup_root_folder, backup_path_template_obj)
+    backup_file = os.path.join(backup_directory, backup_path_template_obj)
 
     if global_settings.backup_test_connectivity is not False:
         task.run(
@@ -81,12 +83,13 @@ def run_backup(  # pylint: disable=too-many-arguments
     backup_obj.backup_last_success_date = task.host.defaults.data["now"]
     backup_obj.backup_config = running_config
     backup_obj.save()
-    logger.log_success(obj, "Successfully backed up device.")
+
+    logger.log_success(obj, "Successfully extracted running configuration from device.")
 
     return Result(host=task.host, result=running_config)
 
 
-def config_backup(job_result, data, backup_root_folder):
+def config_backup(job_result, data):
     """Nornir play to backup configurations."""
     now = datetime.now()
     logger = NornirLogger(__name__, job_result, data.get("debug"))
@@ -130,7 +133,6 @@ def config_backup(job_result, data, backup_root_folder):
                 global_settings=global_settings,
                 remove_regex_dict=remove_regex_dict,
                 replace_regex_dict=replace_regex_dict,
-                backup_root_folder=backup_root_folder,
             )
             logger.log_debug("Completed configuration from devices.")
 
@@ -138,4 +140,4 @@ def config_backup(job_result, data, backup_root_folder):
         logger.log_failure(None, err)
         raise
 
-    logger.log_debug("Completed configuration from devices.")
+    logger.log_debug("Completed configuration backup job for devices.")