+++ nautobot_golden_config/utilities/helper.py
@@ -3,13 +3,17 @@
 
 from jinja2 import exceptions as jinja_errors
 
-from nornir_nautobot.exceptions import NornirNautobotException
-from nautobot.dcim.filters import DeviceFilterSet
+from django import forms
+from django.conf import settings
+
 from nautobot.dcim.models import Device
+from nautobot.dcim.filters import DeviceFilterSet
 from nautobot.utilities.utils import render_jinja2
 
-from nautobot_golden_config import models
+from nornir_nautobot.exceptions import NornirNautobotException
+from nornir_nautobot.utils.logger import NornirLogger
 
+from nautobot_golden_config import models
 
 FIELDS = {
     "platform",
@@ -102,3 +106,60 @@ def render_jinja_template(obj, logger, template):
         )
         logger.log_failure(error_msg)
         raise NornirNautobotException from error
+
+
+def clean_config_settings(repo_type: str, repo_count: int, match_rule: str):
+    """Custom clean for `GoldenConfigSettingFeatureForm`.
+
+    Args:
+        repo_type (str): `intended` or `backup`.
+        repo_count (int): Total number of repos.
+        match_rule (str): Template str provided by user to match repos.
+
+    Raises:
+        ValidationError: Custom Validation on form.
+    """
+    if repo_count > 1:
+        if not match_rule:
+            raise forms.ValidationError(
+                f"If you specify more than one {repo_type} repository, you must provide a {repo_type} repository matching rule template."
+            )
+    elif repo_count == 1 and match_rule:
+        raise forms.ValidationError(
+            f"If you configure only one {repo_type} repository, there is no need to specify the {repo_type} repository matching rule template."
+        )
+
+
+def get_repository_working_dir(
+    repo_type: str,
+    obj: Device,
+    logger: NornirLogger,
+    global_settings: models.GoldenConfigSetting,
+) -> str:
+    """Match the Device to a repository working directory, based on the repository matching rule.
+
+    Assume that the working directory == the slug of the repo.
+
+    Args:
+        repo_type (str): Either `intended` or `backup` repository
+        obj (Device): Django ORM Device object.
+        logger (NornirLogger): Logger object
+        global_settings (models.GoldenConfigSetting): Golden Config global settings.
+
+    Returns:
+        str: The local filesystem working directory corresponding to the repo slug.
+    """
+    match_rule = getattr(global_settings, f"{repo_type}_match_rule")
+
+    if not match_rule:
+        return global_settings.backup_repository.first().filesystem_path
+
+    desired_repository_slug = render_jinja_template(obj, logger, match_rule)
+    matching_repo = getattr(global_settings, f"{repo_type}_repository").filter(slug=desired_repository_slug)
+    if len(matching_repo) == 1:
+        return f"{settings.GIT_ROOT}/{matching_repo[0].slug}"
+    logger.log_failure(
+        obj,
+        f"There is no repository slug matching '{desired_repository_slug}' for device. Verify the matching rule and configured Git repositories.",
+    )
+    return None